pyproject.toml
[project]
name = "cloud-agent"
version = "0.1.0"
description = "Cloud-hosted agent service that executes AI-powered development tasks"
readme = "README.md"
authors = [
    { name = "Dan McAulay", email = "dmcaulay@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "celery[redis]>=5.5.3",
    "e2b-code-interpreter>=2.3.0",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.2.1",
    "redis>=5.2.1",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.0",
]

[project.scripts]
cloud-agent = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "app/tasks/*",
    "app/services/sandbox.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
fail_under = 90
show_missing = true
